Close the Turso v0.8.0-pre.7 gap inventory: corpus refresh, engine parity, extension ports, conformance waves (#61)
## Summary

Continuation of the Ahtola ↔ Turso gap-closure program: analyzes the
remaining feature gaps against the pinned `turso-src` v0.8.0-pre.7
submodule and closes them in seven waves, each gated on the vendored
conformance corpus and the full managed test suite.

**Conformance expected-failures: 185 → 129** (~80 vendored cases closed
across the first five commits; the last two commits are engine
correctness waves pinned by existing tests).

## Waves

### 1. Corpus refresh + engine parity wave (`1c69edc`)
- Vendored 51 new corpus files, refreshed 30 drifted files; the corpus
now byte-matches the pin.
- Added `@var`/matrix grammar to the sqltest runner with a differential
SQLite oracle (`SqltestMatrixOracle`), including `~N` slug dedup for
operator values.
- Fixed ~180 vendored cases: char() argument coercion, `IN
((SELECT...))` membership, i32 bounds for ORDER BY/GROUP BY ordinals,
compound ORDER BY-before-operator errors, qualified `no such table:
main.X` reporting, CHECK-constraint dequote, printf precision cap,
`PRAGMA count_changes`, `ANALYZE sqlite_schema`, and more.
- Implemented `EXPLAIN QUERY PLAN FORMAT=JSON`, built-in
`get_byte`/`set_byte`, and `json_object(*)` star expansion.

### 2. sqlean regexp + time extension ports (`72bd55d`)
- Ported the sqlean regexp family: built-in `regexp` with
`to_text_coerced` semantics, the `X REGEXP Y` operator without
registration, and `regexp_like/substr/replace/capture`.
- Ported the full sqlean time family (~45 functions) with civil-calendar
arithmetic (Hinnant's `days_from_civil`) for BCE years; all 34
`turso/time.sqltest` cases pass.
- Adopted the turso-sqltests subset (regexp, sequence, time, vector,
without_rowid).

### 3. timediff / EXISTS / triggers / windows (`3f3ed13`) — 17 cases
- `timediff()` rewritten with SQLite's directional month arithmetic:
forward month adds clamp day-overflow forward (Jan 31 + 1mo = Mar 2);
negative diffs anchor on the later date and walk backwards (upstream
#8242); sub-day negative intervals keep the sign on the whole interval.
- EXISTS subqueries drop ORDER BY/DISTINCT (dropped terms still
name-resolve; LIMIT/OFFSET stay since OFFSET decides whether a row comes
out).
- Named-window chains: a base defined nowhere errors `no such window`; a
forward reference is silently ignored.
- UPDATE … RETURNING sees the row after the write but before the AFTER
trigger.
- DELETE resets the conflict-policy override for its fired triggers;
temp triggers fire before main triggers in creation order; CTE names
called with args reject with `'cte1' is not a function`.

### 4. json_valid flags cluster (`a4c7e1b`) — 45 cases
- Two-argument `json_valid(X, Y)`: Y is a bitmask (1 = strict RFC 8259
text, 2 = JSON5 text, 4 = shallow JSONB blob, 8 = strict JSONB blob) and
X is valid when any selected check passes; FLAGS coerce like
`sqlite3_value_int`, out-of-range values error with SQLite's exact
message.
- Byte-faithful port of SQLite's `jsonbValidityCheck` drives flags 4/8
and `json_error_position` on JSONB blobs: 32-bit nine-byte header reads,
bare-header requirements for NULL/TRUE/FALSE, digit/hex/canonical-float
payload checks, TEXTJ/TEXT5 escape validation including the
backslash-NUL `strchr` bug-compatibility, line continuations, blind `\u`
conversion after a continuation, and the 1000-deep nesting limit.
- Parser JSON5-construct tracking (`has_json5`) that survives later
standard escapes like upstream's TEXT5/TEXTJ element-type promotion;
JSON5 number mantissa rules with SQLite's exact error positions;
U+2028/U+2029 line continuations; unquoted keys follow SQLite identifier
rules with verbatim `\uXXXX` escapes; `\v` renders through `\u0009`
(≤3.51.1 bug-compat); JSONB TEXT5 payloads store verbatim escapes and
re-read through a dedicated decoder; `subtype()` builtin (74 for JSON
text, 0 otherwise); `%Q`-style bad-path messages; document parsing stops
at the first embedded NUL.
- New `ManagedJsonValidFlagsTests` feature suite.

### 5. Compound ORDER BY + hex-negation overflow (`fc4aae2`) — 6 cases
- Compound SELECT ORDER BY terms match arms structurally via a port of
`exprs_are_equivalent`: function expressions like `upper(b)` match their
output, qualified references match aliased outputs, identifiers compare
case-insensitively, commutative operators match swapped.
- Negated hex literals fold at parse time like SQLite's `codeInteger`:
`-0xNNNN` negates directly while the magnitude fits;
`-0x8000000000000000` is the prepare-time `hex literal too big` error,
while the ALTER TABLE ADD COLUMN backfill still promotes the negation to
REAL 2^63 (upstream #4621); constant DEFAULT expressions are no longer
mis-rejected as non-constant on non-empty tables.

### 6. Partial-index covering + INSERT target alias (`14cecba`) — 5
cases
- A partial index covers a query whose WHERE contains the partial
predicate verbatim: the implied term is dropped from the key-only check
(SQLite `whereLoopAddBtree` pPartIdxWhere), so EQP reports `USING
COVERING INDEX` for both SEARCH and SCAN shapes, and the covering flag
drives the same execution path.
- A rowid-alias column reference under its declared name (`id` for `id
INTEGER PRIMARY KEY`) is always satisfiable from an index entry.
- `INSERT INTO t AS z` parses SQLite's target alias (UPSERT-qualified
references).
- Two stale EQP tests updated to the SQLite-verified COVERING plans.

### 7. Recursive-CTE LIMIT/OFFSET (`7d877e8`) — 12 cases
- A recursive CTE's own LIMIT/OFFSET bounds the recursion like SQLite's
co-routine (`recursive_cte.rs` `init_limit`/`emit_offset`): a literal
LIMIT 0 skips everything including anchor evaluation (its errors never
surface); a negative limit is unlimited; expressions evaluate once up
front; OFFSET rows drop from output but still feed the recursion; the
limit counter stops expansion immediately, combining with the outer row
budget.

## Testing

Every wave ran the vendored conformance corpus for the touched files
plus the full managed suite:

```
pwsh ./scripts/Invoke-ManagedTestSuite.ps1 -Framework net10.0 -MinimumExecutedTests 2500
```

Final state: **7,757 passed / 0 failed** on net10.0. Remaining
expected-failures (129) are documented in
`managed-sqltest-expected-failures.txt`: recursive-cte
priority-queue/nested-CTE rules, unnest-correlated, window frame
matrices, and intentional extensions (DELETE/UPDATE LIMIT, temp-view
restrictions) that Ahtola deliberately keeps beyond the pinned Turso
baseline.

Refs #263

+++ src/Ahtola.Tests/WindowAndReturningTests.cs
@@ -258,6 +258,8 @@ public class WindowAndReturningTests
             END;
             """);
 
+        // SQLite's RETURNING subquery reads the target table after the write but
+        // before the AFTER trigger (returning.sqltest), so both projections see 20.
         ReadRows(
                 connection,
                 """
@@ -266,7 +268,7 @@ public class WindowAndReturningTests
                 """)
             .Should()
             .ContainSingle()
-            .Which.Should().Equal(SqlValue.Integer(120), SqlValue.Integer(120));
+            .Which.Should().Equal(SqlValue.Integer(20), SqlValue.Integer(20));
     }
 
     [Test]
